+++ CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Add `PlaceManager.errorPage(BiFunction<Place, String, Page>)`
+
 ## [2.2.1] - 2025-09-22
 
 ### Fixed